+++ bin/in-container/should-be-the-same.sh
@@ -35,6 +35,7 @@ if [ $# != 0 ]; then
 fi
 
 confirm_same check-consistent-dev-env-version.sh
+confirm_same create-dummy-docker-container.sh
 confirm_same repo.sh
 confirm_same repo-root-dir.sh